+++ src/directives/componentRenderer/messageRenderer.directive.ts
@@ -1,4 +1,4 @@
-import {ComponentRef, Directive, Input, OnChanges, SimpleChanges, ViewContainerRef, EventEmitter, Output, ComponentFactoryResolver, OnDestroy, Inject, Optional, Type} from '@angular/core';
+import {ComponentRef, Directive, Input, OnChanges, SimpleChanges, ViewContainerRef, EventEmitter, Output, OnDestroy, Inject, Optional, Type} from '@angular/core';
 import {Notification} from '@anglr/common';
 import {nameof} from '@jscrpt/common';
 import {Subscription} from 'rxjs';
@@ -20,26 +20,26 @@ export class MessageRendererDirective implements OnChanges, OnDestroy
     /**
      * Created component reference
      */
-    protected _componentRef: ComponentRef<NotificationMessage>|null = null;
+    protected componentRef: ComponentRef<NotificationMessage>|null = null;
 
     /**
      * Subscription for listening on closed message
      */
-    protected _closedSubscription: Subscription;
+    protected closedSubscription: Subscription|undefined|null;
 
     //######################### public properties - inputs #########################
 
     /**
      * Instance of notification message that should be rendered
      */
-    @Input('messageRenderer')
-    public notification: Notification;
+    @Input({alias: 'messageRenderer', required: true})
+    public notification!: Notification;
 
     /**
      * Options passed to notifications component
      */
-    @Input()
-    public notificationsOptions: NotificationsOptions;
+    @Input({required: true})
+    public notificationsOptions!: NotificationsOptions;
 
     //######################### public properties - outputs #########################
 
@@ -56,17 +56,17 @@ export class MessageRendererDirective implements OnChanges, OnDestroy
      */
     protected get component(): NotificationMessage|null
     {
-        if(!this._componentRef)
+        if(!this.componentRef)
         {
             return null;
         }
 
-        return this._componentRef.instance;
+        return this.componentRef.instance;
     }
 
     //######################### constructor #########################
-    constructor(protected _viewContainerRef: ViewContainerRef,
-                @Inject(NOTIFICATION_MESSAGE_SERVICE) @Optional() protected _notificationMessageService?: NotificationMessageService)
+    constructor(protected viewContainerRef: ViewContainerRef,
+                @Inject(NOTIFICATION_MESSAGE_SERVICE) @Optional() protected notificationMessageService?: NotificationMessageService)
     {
     }
 
@@ -77,32 +77,37 @@ export class MessageRendererDirective implements OnChanges, OnDestroy
      */
     public ngOnChanges(changes: SimpleChanges): void
     {
-        this._viewContainerRef.clear();
+        this.viewContainerRef.clear();
 
         if(nameof<MessageRendererDirective>('notification') in changes && changes[nameof<MessageRendererDirective>('notification')].currentValue)
         {
-            const injector = this._viewContainerRef.injector;
+            const injector = this.viewContainerRef.injector;
             let notificationMessageType: Type<NotificationMessage>;
 
-            if(this._notificationMessageService)
+            if(this.notificationMessageService)
             {
-                notificationMessageType = this._notificationMessageService.getNotificationMessageComponent(this.notification.severity);
+                notificationMessageType = this.notificationMessageService.getNotificationMessageComponent(this.notification.severity);
             }
             else
             {
                 notificationMessageType = this.notificationsOptions.getNotificationMessageComponent(this.notification.severity);
             }
 
-            const componentFactoryResolver: ComponentFactoryResolver = injector.get(ComponentFactoryResolver);
-            const componentFactory = componentFactoryResolver.resolveComponentFactory(notificationMessageType);
-            this._componentRef = this._viewContainerRef.createComponent(componentFactory, this._viewContainerRef.length, injector);
+            this.componentRef = this.viewContainerRef.createComponent(notificationMessageType, 
+                                                                      {
+                                                                          injector,
+                                                                      });
 
             this.ngOnDestroy();
 
-            this._closedSubscription = this.component.closed.subscribe(this._closedEmit);
-            this.component.item = this.notification;
-            this.component.options = this.notificationsOptions.messageOptions ?? {};
-            this.component.invalidateVisuals();
+            this.closedSubscription = this.component?.closed.subscribe(this.closedEmit);
+
+            if(this.component)
+            {
+                this.component.item = this.notification;
+                this.component.options = this.notificationsOptions.messageOptions ?? {};
+                this.component.invalidateVisuals();
+            }
         }
     }
 
@@ -113,8 +118,8 @@ export class MessageRendererDirective implements OnChanges, OnDestroy
      */
     public ngOnDestroy(): void
     {
-        this._closedSubscription?.unsubscribe();
-        this._closedSubscription = null;
+        this.closedSubscription?.unsubscribe();
+        this.closedSubscription = null;
     }
 
     //######################### protected methods #########################
@@ -123,7 +128,7 @@ export class MessageRendererDirective implements OnChanges, OnDestroy
      * Called when notification was closed by user
      * @param notification - Notification that was closed
      */
-    protected _closedEmit = (notification: Notification): void =>
+    protected closedEmit = (notification: Notification): void =>
     {
         this.closed.emit(notification);
     }